apply refactoring to enable visualization and add some statistics for plots to output of ga
