+++ main.py
@@ -1,52 +1,5 @@
 from datasets.get_dataset import get_dataset
-from genetic.evaluation import evaluate
-from genetic.population import apply_crossover, apply_mutation, generate_population
-from typing import List
-
-from tqdm import tqdm
-import copy
-
-from genetic.chromosome import Chromosome
-
-import numpy as np
-
-
-def genetic_algorithm(configuration, dataset):
-    population = generate_population(configuration["population_size"])
-    total_pipelines_over_epochs = []
-    seen_pipelines = []
-
-    for epoch in range(configuration["epochs"]):
-        population, old_scores = run_epoch(population, dataset, configuration)
-        print(
-            f"Epoch {epoch} mean: {np.mean(old_scores):.02f} best: {old_scores[0]:.03f}"
-        )
-
-        for pl in population:
-            if str(pl) not in seen_pipelines:
-                seen_pipelines.append(str(pl))
-
-        total_pipelines_over_epochs.append(len(seen_pipelines))
-
-    return population[0], total_pipelines_over_epochs
-
-
-def run_epoch(population, dataset, configuration) -> (List[Chromosome], List[float]):
-    scores = []
-    for p in tqdm(population):
-        scores.append(evaluate(p, dataset.copy(), configuration))
-
-    sorted_population = [x for _, x in sorted(zip(scores, population))]
-    print(sorted_population[0])
-
-    elite_num = int(0.2 * len(sorted_population))
-    elites = [copy.deepcopy(x) for x in sorted_population[:elite_num]]
-
-    new_population = apply_mutation(sorted_population, configuration["mutation_prob"])
-    new_population = apply_crossover(new_population, configuration["crossover_prob"])
-
-    return elites + new_population[0 : len(population) - elite_num], sorted(scores)
-
+from genetic.algorithm import genetic_algorithm
 
 if __name__ == "__main__":
     configuration = {
@@ -60,7 +13,7 @@ if __name__ == "__main__":
 
     dataset = get_dataset("electricity", size=1000)
 
-    best_chromosome, num_pipelines_over_epochs = genetic_algorithm(
+    best_chromosome, num_pipelines_over_epochs, mean_per_epoch = genetic_algorithm(
         configuration, dataset
     )
     print(num_pipelines_over_epochs)
